ISSUE #475 refactored dwg to have it's own file processor
